refactor: integrate study session start into page
